refactor: use single obfuscated id in export route
