+++ package.json
@@ -16,6 +16,7 @@
     "pumpify": "^2.0.0",
     "rimraf": "^3.0.0",
     "routes": "^2.1.0",
+    "run-parallel": "^1.1.9",
     "safe-fs-blob-store": "^1.0.6",
     "through2": "^3.0.1"
   }
